Create `/components`, move `Header`
- Components files will be added to the new folder
- Avoids creating routes for every single components, avoids reference errors, enables better organization
