+++ pages/index.js
@@ -2,7 +2,7 @@ import Head from "next/head";
 import Image from "next/image";
 import styles from "../styles/Home.module.css";
 
-import Header from "./Header";
+import Header from "./components/Header";
 
 export default function Home() {
 	const data = {
